# [WIP] [EXPERIMENT] [DONT MERGE] Speed up "get one" and "get all" / query MongoDB (mongoengine) operations