## Background, Details

There is a known issue with some mongoengine operations (especially retrievals and to lesser extend, also insertions) being very slow when working with large documents (https://github.com/MongoEngine/mongoengine/issues/1230).

The primary reason for that is massive overhead mongoengine adds when converting native pymongo representation to the document class instance and vice versa.

There are multiple approaches on how to tackle that, including switching to something like pymodm which is much more efficient than mongoengine.

Most of those libraries claim very similar and compatible API to the mongoengine, but this doesn't help us at all. 

Our DB later abstraction is based heavily around mongoengine and we utilize a lot of low level mongoengine functionality and primitives (and also extend it in various places).

This means that switching the ORM library we use is pretty much impossible at the moment. It would simply require too much time and resources and it would likely also result in a lot of hard to catch / edge case bugs.

## Proposed Solution

While profiling and measuring hot code paths and trying to see which places we could optimize (switch directly to pymongo and avoid mongoengine layer all together), I found out that calling ``as_pymongo()`` on mongoengine QuerySet object results in massive performance improvements.

The nice thing is that it's much easier to make that change compatible with out existing code and DB layer compared to fully switching an ORM layer which would require rewriting very big chunks of the code.

The reason for that is that this method avoids very expensive mongoengine conversion and dereferrencing layer which we don't need anyway. It simply returns raw pymongo result aka dictionary.

In this proposed implementation, I updated our ``get()`` and ``query()`` method to utilize ``as_pymongo()`` method and manually convert those raw dictionaries to our database model class instances.

It turns out that approach is much more efficient.

As part of that changes, I also need to update various code to utilize ``query()`` in a manner that it expects that method to return actual DB model class instances and not a ``QuerySet`` object.

In fact, we already did that for a lot of the code in the past. If you check, 90% of the code which calls ``query()`` already expects it to return actual DB model objects and not a query set, so the change shouldn't be that big.

In places where there is not possible and where we need access to raw QuerySet (there should be very few of those), we can use ``raw_query()`` method directly.

Keep in mind that this approach is still not 100% ideal. In ideal world, we would avoid instantiating DB class all together and work directly with raw pymongo dictionaries (at least in critical code paths), but that change would be much harder and more involved.

## Profiling Data

Here is some profiling data when retrieving a single large object with and without utilizing as_pymongo.

```bash
Retrieve large execution with as_pymongo=False
99.9 percentile: 0.011759551286697398s
99 percentile: 0.011098940372467038s
95 percentile: 0.009706759452819822s
90 percentile: 0.011098940372467038s

Retrieve large execution with as_pymongo=True 
99.9 percentile: 0.005786781072616579s
99 percentile: 0.005684897899627685s
95 percentile: 0.005177915096282959s
90 percentile: 0.005684897899627685s
```

As you can see, the difference is substantial (aka as_pymongo + manual model object instantiation around 50% faster).

## TODO

Overall, I'm quite optimistic I can get all that change finished and call the tests passing compared to many other much more involved and complex approach (I already have a huge chunk of tests passing locally).

- [ ] Update affected code and make sure all the tests pass