@Kami 2019-12-19
I'm also looking if I can get rid of auto conversion / dereferencing in the base document DB class.

This only needs to happen where ``EmbeddedDocumentField`` field type is used. We only use that field type in a couple of places, yet dereferrencing adds huge overhead.

There are multiple ways to get rid of auto conversion - probably the easiest / best way is to change those field types to simple Dict / EscapedDynamicField fields.

Another option is to explicitly try to de-reference only those fields which use that field type in a custom class constructor.

I will look into both approaches, second one would probably require less changed.