@Kami 2019-12-19
And here is the WIP approach with manual de-referencing I'm testing - a7582ae6a.

It speeds those two methods up by 20-30% so it's, imo, well worth it (and we don't really have any better option anyway).

---

Another problematic and very slow part is using ``EscapedDictField`` and ``EscapedDyanmicField``.

It means we need to unescape the values when we retrieve those models from the database and that's a very expensive and slow process (especially on large and nested dicts - I tested it with and without escaping and without it's around 5x faster which in absolute numbers is around 0.2s vs 0.02s.).

It seems like that method of escaping is not recommend anymore anyway - https://jira.mongodb.org/browse/DOCS-9311.

Sadly we don't have much choice here, we still need to support escaping (at least for a while) for backward compatibility reasons.