@Kami 2019-12-19
First step on optimizing those escaped fields would be changing fields which are controlled by us and can't contain ``$`` and ``.`` characters to be regular dict fields.

@m4dcoder Can all those fields contain special characters ($, .) - https://gist.github.com/Kami/04718da61cab79b02e2a2a19c3f4f397?

Since as per my comment, using that field type adds massive overhead on object retrieval + insertion time.

Another alternative for fields such as ``result`` would be to store result as JSON serialized string. I still need to do some more benchmarking, but I believe that should be faster than escaping and unescaping very large result dictionaries.