@Kami 2019-12-19
I did some retrieval benchmarking to see if we could indeed utilize a special field type which stores large dictionaries / values as JSON serialized strings.

EscapedDynamicField with character unescaping on retrieval:

```bash
99.9 percentile: 0.11631401681900028s
99 percentile: 0.1144586706161499s
95 percentile: 0.0900427460670471s
90 percentile: 0.1144586706161499s
```

Using JSON field type where unescapping is not necessary:

```bash
99.9 percentile: 0.06011564707756043s
99 percentile: 0.059931752681732176s
95 percentile: 0.0596500039100647s
90 percentile: 0.059931752681732176s
```

Only "downside" of using a JSONDictField is that we can't query on actual dictionary values, but that's not something we are doing anyway and for all purposes, ``result`` field an opaque string / dictionary for us.

Now the main problem is how to implement this change in a backward compatible manner since we don't really have object / model versioning in place.

One approach is to do something like this and handle conversion from old format to a new one "lazily" on initial retrieval / write of the existing objects:

1. Add new ``model_converted_to_new_json_field_type`` field or similar to all the affected models and default it to ``True`` for new models on insertion and ``False`` for existing models on retrieval.

2. When retrieving an existing object, if that field is not set, assume we are working with the old data format, unescape the dictionary and on save, store it in MongoDB in a new format and set ``model_converted_to_new_json_field_type`` field to ``True``.

3. When writing a new object we can just assume and use new format (same as we are retrieving those objects from a database)

This applies to existing installations which upgrade to a new version aka backward compatibility layer. If we wouldn't need to support that (but we don't really have a choice), the change would obviously be much more straight forward.

Another option would be to add a data migration script, but this could potentially cause a long down time for installations with a lot of existing database models, so I think "lazy" approach is the way to go.

I believe that with that change + other changes in this PR we would get retrieval performance down to a reasonable level for the foreseeable future.